fix comments directory
