+++ src/features/comments/episodes/hooks/useCommentsEpisode.ts
@@ -1,43 +1,38 @@
 import { useQueryClient, useIsFetching } from "@tanstack/react-query";
 import { useRouter } from "next/router";
-import { useMemo } from "react";
+import { useMemo, useCallback } from "react";
 import { useInView } from "react-intersection-observer";
-import { useInfiniteCommentsWork } from "src/features/comments/api/useInfiniteCommentsWork";
-
-import {
-  InfiniteCommentEpisode,
-  InfiniteCommentWork,
-} from "src/features/comments/types";
-import { getInitialPageParam } from "src/features/comments/utils";
-import { getLatestWorkComments } from "src/features/comments/utils/getLatestComments";
+import { useInfiniteCommentsEpisode } from "src/features/comments/api/useInfiniteCommentsEpisode";
+import { InfiniteCommentEpisode } from "src/features/comments/common/types";
+import { getInitialPageParam } from "src/features/comments/common/utils";
+import { getLatestEpisodeComments } from "src/features/comments/common/utils/getLatestComments";
 import { GetLatestEpisodeCommentQuery } from "src/gql/graphql";
 
-export const useCommentsWork = (work_id: number) => {
+export const useCommentsEpisode = (episode_id: string) => {
   const router = useRouter();
   const order = router.query.order === "popular" ? "popular" : "new";
   const { data, isFetchingNextPage, fetchNextPage, hasNextPage, isPending } =
-    useInfiniteCommentsWork(work_id, order);
-
+    useInfiniteCommentsEpisode(episode_id, order);
   const queryClient = useQueryClient();
   const isLatestFetching = useIsFetching({
-    queryKey: ["latestComments", { work_id }],
+    queryKey: ["latestComments", { episode_id }],
   });
 
   const { ref } = useInView({
     root: null,
     onChange: (inView) => {
       if (isFetchingNextPage || !inView) return;
+
       fetchNextPage();
     },
     threshold: 1.0,
   });
 
-  const refetchHandler = async () => {
-    if (!work_id || isFetchingNextPage) return;
-
-    const prevData = queryClient.getQueryData<InfiniteCommentWork>([
+  const refetchHandler = useCallback(async () => {
+    if (!episode_id || isFetchingNextPage) return;
+    const prevData = queryClient.getQueryData<InfiniteCommentEpisode>([
       "comments",
-      { work_id, order },
+      { episode_id, order },
     ]);
 
     if (!prevData) return;
@@ -47,11 +42,11 @@ export const useCommentsWork = (work_id: number) => {
     const latestData =
       await queryClient.fetchQuery<GetLatestEpisodeCommentQuery>({
         queryFn: () =>
-          getLatestWorkComments({
-            work_id,
+          getLatestEpisodeComments({
+            episode_id,
             cursor,
           }),
-        queryKey: ["latestComments", { work_id }],
+        queryKey: ["latestComments", { episode_id }],
         gcTime: 0,
       });
 
@@ -64,7 +59,7 @@ export const useCommentsWork = (work_id: number) => {
     if (newPagesArray.length === 0) return;
 
     queryClient.setQueryData<InfiniteCommentEpisode>(
-      ["comments", { work_id, order }],
+      ["comments", { episode_id, order }],
       {
         pageParams: [getInitialPageParam(), ...prevData.pageParams],
         pages: [
@@ -75,10 +70,11 @@ export const useCommentsWork = (work_id: number) => {
         ],
       }
     );
+
     queryClient.invalidateQueries({
       queryKey: ["replies"],
     });
-  };
+  }, [episode_id, isFetchingNextPage, order, queryClient]);
 
   const memoData = useMemo(() => {
     if (!data?.pages) return [];
