+++ src/features/comments/works/hooks/useCommentsWork.ts
@@ -1,38 +1,43 @@
 import { useQueryClient, useIsFetching } from "@tanstack/react-query";
 import { useRouter } from "next/router";
-import { useMemo, useCallback } from "react";
+import { useMemo } from "react";
 import { useInView } from "react-intersection-observer";
-import { useInfiniteCommentsEpisode } from "src/features/comments/api/useInfiniteCommentsEpisode";
-import { InfiniteCommentEpisode } from "src/features/comments/types";
-import { getInitialPageParam } from "src/features/comments/utils";
-import { getLatestEpisodeComments } from "src/features/comments/utils/getLatestComments";
+import { useInfiniteCommentsWork } from "src/features/comments/api/useInfiniteCommentsWork";
+
+import {
+  InfiniteCommentEpisode,
+  InfiniteCommentWork,
+} from "src/features/comments/common/types";
+import { getInitialPageParam } from "src/features/comments/common/utils";
+import { getLatestWorkComments } from "src/features/comments/common/utils/getLatestComments";
 import { GetLatestEpisodeCommentQuery } from "src/gql/graphql";
 
-export const useCommentsEpisode = (episode_id: string) => {
+export const useCommentsWork = (work_id: number) => {
   const router = useRouter();
   const order = router.query.order === "popular" ? "popular" : "new";
   const { data, isFetchingNextPage, fetchNextPage, hasNextPage, isPending } =
-    useInfiniteCommentsEpisode(episode_id, order);
+    useInfiniteCommentsWork(work_id, order);
+
   const queryClient = useQueryClient();
   const isLatestFetching = useIsFetching({
-    queryKey: ["latestComments", { episode_id }],
+    queryKey: ["latestComments", { work_id }],
   });
 
   const { ref } = useInView({
     root: null,
     onChange: (inView) => {
       if (isFetchingNextPage || !inView) return;
-
       fetchNextPage();
     },
     threshold: 1.0,
   });
 
-  const refetchHandler = useCallback(async () => {
-    if (!episode_id || isFetchingNextPage) return;
-    const prevData = queryClient.getQueryData<InfiniteCommentEpisode>([
+  const refetchHandler = async () => {
+    if (!work_id || isFetchingNextPage) return;
+
+    const prevData = queryClient.getQueryData<InfiniteCommentWork>([
       "comments",
-      { episode_id, order },
+      { work_id, order },
     ]);
 
     if (!prevData) return;
@@ -42,11 +47,11 @@ export const useCommentsEpisode = (episode_id: string) => {
     const latestData =
       await queryClient.fetchQuery<GetLatestEpisodeCommentQuery>({
         queryFn: () =>
-          getLatestEpisodeComments({
-            episode_id,
+          getLatestWorkComments({
+            work_id,
             cursor,
           }),
-        queryKey: ["latestComments", { episode_id }],
+        queryKey: ["latestComments", { work_id }],
         gcTime: 0,
       });
 
@@ -59,7 +64,7 @@ export const useCommentsEpisode = (episode_id: string) => {
     if (newPagesArray.length === 0) return;
 
     queryClient.setQueryData<InfiniteCommentEpisode>(
-      ["comments", { episode_id, order }],
+      ["comments", { work_id, order }],
       {
         pageParams: [getInitialPageParam(), ...prevData.pageParams],
         pages: [
@@ -70,11 +75,10 @@ export const useCommentsEpisode = (episode_id: string) => {
         ],
       }
     );
-
     queryClient.invalidateQueries({
       queryKey: ["replies"],
     });
-  }, [episode_id, isFetchingNextPage, order, queryClient]);
+  };
 
   const memoData = useMemo(() => {
     if (!data?.pages) return [];
