+++ src/features/episodes/components/index.tsx
@@ -9,8 +9,8 @@ import { ErrorMessage } from "src/components/Error/items/ErrorMessage";
 import { TimerModal } from "src/components/Modal/Timer";
 import { EpisodeChatInput } from "src/features/chats/episodes/components/EpisodeChatInput";
 import { EpisodeChats } from "src/features/chats/episodes/components/EpisodeChats";
-import { EpisodeCommentInput } from "src/features/comments/components/EpisodeCommentInput";
-import { EpisodeComments } from "src/features/comments/components/EpisodeComments";
+import { EpisodeCommentInput } from "src/features/comments/episodes/components/EpisodeCommentInput";
+import { EpisodeComments } from "src/features/comments/episodes/components/EpisodeComments";
 import { useEpisode } from "src/features/episodes/hooks/useEpisode";
 import { getIsStatus } from "src/features/timer/utils/getIsStatus";
 import { GraphQLError } from "src/libs/error";
