+++ src/features/episodes/mocks/msw.ts
@@ -1,6 +1,6 @@
 import { graphql } from "msw";
 import { handleEpisodeChat } from "src/features/chats/common/mocks/msw";
-import { handlers as commentHandlers } from "src/features/comments/mocks/msw";
+import { handlers as commentHandlers } from "src/features/comments/common/mocks/msw";
 import { episodeData } from "src/features/episodes/mocks/fixture";
 
 export const handleEpisode = (status?: number) => {
