+++ src/features/live/components/index.tsx
@@ -6,8 +6,8 @@ import { Header } from "src/components/dynamic/live/header";
 import { Nav } from "src/components/dynamic/live/nav";
 import { Loader } from "src/components/Elements/Loader";
 import { ErrorMessage } from "src/components/Error/items/ErrorMessage";
-import { EpisodeCommentInput } from "src/features/comments/components/EpisodeCommentInput";
-import { EpisodeComments } from "src/features/comments/components/EpisodeComments";
+import { EpisodeCommentInput } from "src/features/comments/episodes/components/EpisodeCommentInput";
+import { EpisodeComments } from "src/features/comments/episodes/components/EpisodeComments";
 import { FinishLive } from "src/features/live/components/Finish";
 import { LiveChatInput } from "src/features/live/components/LiveChatInput";
 import { LiveChats } from "src/features/live/components/LiveChats";
