+++ src/features/works/components/PlayWork.tsx
@@ -9,8 +9,8 @@ import { ErrorMessage } from "src/components/Error/items/ErrorMessage";
 import { TimerModal } from "src/components/Modal/Timer";
 import { WorkChatInput } from "src/features/chats/works/components/WorkChatInput";
 import { WorkChats } from "src/features/chats/works/components/WorkChats";
-import { WorkCommentInput } from "src/features/comments/components/WorkCommentInput";
-import { WorkComments } from "src/features/comments/components/WorkComments";
+import { WorkCommentInput } from "src/features/comments/works/components/WorkCommentInput";
+import { WorkComments } from "src/features/comments/works/components/WorkComments";
 import { usePlayWork } from "src/features/works/hooks/usePlayWork";
 import { NotFoundError } from "src/libs/error";
 import { DetailTitle } from "src/libs/meta/OnlyTitle";
